refactor: TimerFocus에 있는 State들 커스텀 훅으로 분리
